slash changes
