+++ src/tasks/minions/bso/kalphiteKingActivity.ts
@@ -206,7 +206,7 @@ export default class extends Task {
 					  }, you died ${deaths[userID] ?? 0} times. Your Kalphite King KC is now ${
 							leaderUser.settings.get(UserSettings.MonsterScores)[KalphiteKingMonster.id] ?? 0
 					  }.\n\n${soloXP}`,
-				['kk', ['solo'], true],
+				['k', { name: 'Kalphite King' }, true],
 				image!,
 				data,
 				soloItemsAdded
